pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = ""
authors = ["Thiago Almeida <thiagolsa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-decouple = "^3.3"
python-telegram-bot = "^12.2"
tinydb = "^3.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
